Use getConfigForFile to handle configurations
Move to using `CLIEngine#getConfigForFile` to handle determination of
the configuration for a file instead of our attempts to replicate this
behavior.

Note that if a user has a configuration in their home directory, and
nowhere more specific for a file, previously we were counting this as
"no configuration" for the purposes of the "Disable when no config"
option. After this change we are doing the same thing ESLint does: Using
this configuration. If this isn't what a user desires, they should move
away from a "global config".